requirements.txt
dnslib
dnspython